backend/requirements.txt
absl-py==2.2.0
annotated-types==0.6.0
anyio==4.7.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.12.3
Brotli==1.1.0
cached-property==1.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dnspython==1.16.0
exceptiongroup==1.3.0
executing==2.2.0
fake-useragent==2.2.0
fastapi==0.112.2
flatbuffers==25.2.10
fonttools==4.58.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.67.1
h11==0.16.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.2
kiwisolver==1.4.8
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
motor==3.7.1
munkres==1.1.4
namex==0.0.9
nest_asyncio==1.6.0
numpy>=1.26.0,<2.1.0
opt_einsum==3.4.0
optree==0.15.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pip==25.1.1
platformdirs==4.3.8
playwright==1.52.0
prompt_toolkit==3.0.51
protobuf==5.28.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyee==13.0.0
Pygments==2.19.1
pymongo==4.13.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.4.0
RapidFuzz==3.13.0
requests==2.32.3
rich==14.0.0
scikeras==0.13.0
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.33.0
setuptools==80.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack_data==0.6.3
starlette==0.38.2
tensorboard==2.18.0
tensorboard_data_server==0.7.0
tensorflow==2.18.0
tensorflow_estimator==2.15.0
termcolor==3.1.0
threadpoolctl==3.6.0
tornado==6.4.2
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
unicodedata2==16.0.0
urllib3==2.4.0
uvicorn==0.34.3
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
zstandard==0.23.0
